Extract SchoolSelector to its own module
